Update DESCRIPTION
The CRAN policy asks for minimal package size, which includes installed
size.

Please re-read 'Writing R Extensions' §1.1.6.  That LazyDataCompression
is not used gives a warning in R-devel.  Using the
CheckLazyDataCompression() function in that section gives (database
sizes in KB)

  gzip bzip2    xz
  5510  4695  4038

So there are substantial gains by adding

LazyDataCompression: xz

to the DESCRIPTION file: please do so.